refactor: reorganiza pacotes de casos de uso e DTOs
- Move interfaces de casos de uso de `api.usecases` para `usecases`.
- Move implementações internas de casos de uso para `usecases.impl`.
- Move DTOs de `api.usecases.dto` para `usecases.dto`.
- Adiciona a interface pública `UserPublicApi` em `users.api` e sua implementação correspondente.

Refs: GH-1
