+++ src/main/java/dev/ghonda/project/management/users/api/UserPublicApi.java
@@ -1,0 +1,12 @@
+package dev.ghonda.project.management.users.api;
+
+import dev.ghonda.project.management.users.domain.User;
+
+import java.util.Optional;
+
+@FunctionalInterface
+public interface UserPublicApi {
+
+    Optional<User> findUserByUsername(String username);
+
+}
