--- src/main/java/dev/ghonda/project/management/users/api/usecases/package-info.java
@@ -1,4 +1,0 @@
-@NamedInterface("usecases")
-package dev.ghonda.project.management.users.api.usecases;
-
-import org.springframework.modulith.NamedInterface;
