+++ src/main/java/dev/ghonda/project/management/users/usecases/RegisterUserUseCase.java
@@ -1,7 +1,7 @@
-package dev.ghonda.project.management.users.api.usecases;
+package dev.ghonda.project.management.users.usecases;
 
 import dev.ghonda.project.management.shared.dto.Resource;
-import dev.ghonda.project.management.users.api.usecases.dto.RegisterUserPayload;
+import dev.ghonda.project.management.users.usecases.dto.RegisterUserPayload;
 
 @FunctionalInterface
 public interface RegisterUserUseCase {
