+++ src/main/java/dev/ghonda/project/management/users/usecases/SearchUserByIdUseCase.java
@@ -1,0 +1,10 @@
+package dev.ghonda.project.management.users.usecases;
+
+import dev.ghonda.project.management.users.usecases.dto.UserDetailPayload;
+
+@FunctionalInterface
+public interface SearchUserByIdUseCase {
+
+    UserDetailPayload execute(Long userId);
+
+}
