+++ src/main/java/dev/ghonda/project/management/users/usecases/SearchUsersUseCase.java
@@ -1,6 +1,6 @@
-package dev.ghonda.project.management.users.api.usecases;
+package dev.ghonda.project.management.users.usecases;
 
-import dev.ghonda.project.management.users.api.usecases.dto.UserDetailPayload;
+import dev.ghonda.project.management.users.usecases.dto.UserDetailPayload;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
