+++ src/main/java/dev/ghonda/project/management/users/usecases/UpdateUserUseCase.java
@@ -1,0 +1,11 @@
+package dev.ghonda.project.management.users.usecases;
+
+import dev.ghonda.project.management.users.usecases.dto.UpdateUserPayload;
+import dev.ghonda.project.management.users.usecases.dto.UserDetailPayload;
+
+@FunctionalInterface
+public interface UpdateUserUseCase {
+
+    UserDetailPayload execute(Long userId, UpdateUserPayload payload);
+
+}
