+++ src/main/java/dev/ghonda/project/management/users/usecases/impl/RegisterUserUseCaseImpl.java
@@ -1,4 +1,4 @@
-package dev.ghonda.project.management.users.internal.usecases;
+package dev.ghonda.project.management.users.usecases.impl;
 
 import dev.ghonda.project.management.shared.annotations.UseCase;
 import dev.ghonda.project.management.shared.dto.Resource;
@@ -7,8 +7,8 @@ import dev.ghonda.project.management.shared.validators.ValidatorService;
 import dev.ghonda.project.management.users.domain.Role;
 import dev.ghonda.project.management.users.domain.User;
 import dev.ghonda.project.management.users.domain.UserRepository;
-import dev.ghonda.project.management.users.api.usecases.RegisterUserUseCase;
-import dev.ghonda.project.management.users.api.usecases.dto.RegisterUserPayload;
+import dev.ghonda.project.management.users.usecases.RegisterUserUseCase;
+import dev.ghonda.project.management.users.usecases.dto.RegisterUserPayload;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
