+++ src/main/java/dev/ghonda/project/management/users/usecases/impl/SearchUserByIdUseCaseImpl.java
@@ -1,10 +1,10 @@
-package dev.ghonda.project.management.users.internal.usecases;
+package dev.ghonda.project.management.users.usecases.impl;
 
 import dev.ghonda.project.management.shared.annotations.UseCase;
 import dev.ghonda.project.management.shared.exceptions.ResourceNotFoundException;
 import dev.ghonda.project.management.users.domain.UserRepository;
-import dev.ghonda.project.management.users.api.usecases.dto.UserDetailPayload;
-import dev.ghonda.project.management.users.api.usecases.SearchUserByIdUseCase;
+import dev.ghonda.project.management.users.usecases.dto.UserDetailPayload;
+import dev.ghonda.project.management.users.usecases.SearchUserByIdUseCase;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
