+++ src/main/java/dev/ghonda/project/management/users/usecases/impl/SearchUsersUseCaseImpl.java
@@ -1,9 +1,9 @@
-package dev.ghonda.project.management.users.internal.usecases;
+package dev.ghonda.project.management.users.usecases.impl;
 
 import dev.ghonda.project.management.shared.annotations.UseCase;
 import dev.ghonda.project.management.users.domain.UserRepository;
-import dev.ghonda.project.management.users.api.usecases.dto.UserDetailPayload;
-import dev.ghonda.project.management.users.api.usecases.SearchUsersUseCase;
+import dev.ghonda.project.management.users.usecases.dto.UserDetailPayload;
+import dev.ghonda.project.management.users.usecases.SearchUsersUseCase;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
