+++ src/main/java/dev/ghonda/project/management/users/usecases/impl/UpdateUserUseCaseImpl.java
@@ -1,13 +1,13 @@
-package dev.ghonda.project.management.users.internal.usecases;
+package dev.ghonda.project.management.users.usecases.impl;
 
 import dev.ghonda.project.management.shared.annotations.UseCase;
 import dev.ghonda.project.management.shared.exceptions.ResourceNotFoundException;
 import dev.ghonda.project.management.shared.validators.ValidatorService;
 import dev.ghonda.project.management.users.domain.Role;
 import dev.ghonda.project.management.users.domain.UserRepository;
-import dev.ghonda.project.management.users.api.usecases.dto.UpdateUserPayload;
-import dev.ghonda.project.management.users.api.usecases.dto.UserDetailPayload;
-import dev.ghonda.project.management.users.api.usecases.UpdateUserUseCase;
+import dev.ghonda.project.management.users.usecases.dto.UpdateUserPayload;
+import dev.ghonda.project.management.users.usecases.dto.UserDetailPayload;
+import dev.ghonda.project.management.users.usecases.UpdateUserUseCase;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.transaction.annotation.Transactional;
